@@ -4,29 +4,29 @@ import '../core.dart';
 
 class RenderZToBoxAdapter extends RenderZBox
     with RenderObjectWithChildMixin<RenderBox> {
-  double _width;
+  double? _width;
 
-  double get width => _width;
+  double? get width => _width;
 
-  set width(double value) {
+  set width(double? value) {
     if (_width == value) return;
     _width = value;
     markNeedsLayout();
   }
 
-  double _height;
+  double? _height;
 
-  double get height => _height;
+  double? get height => _height;
 
-  set height(double value) {
+  set height(double? value) {
     if (_height == value) return;
     _height = value;
     markNeedsLayout();
   }
 
   RenderZToBoxAdapter({
-    double width,
-    double height,
+    double? width,
+    double? height,
   })  : _width = width,
         _height = height;
 
@@ -35,17 +35,17 @@ class RenderZToBoxAdapter extends RenderZBox
 
   // bool get isRepaintBoundary => true;
 
-  List<ZPathCommand> transformedPath;
+  late List<ZPathCommand> transformedPath;
 
   @override
   void performLayout() {
     final ZParentData anchorParentData = parentData as ZParentData;
-    child.layout(BoxConstraints.expand(height: height, width: width),
+    child!.layout(BoxConstraints.expand(height: height, width: width),
         parentUsesSize: false);
     size = constraints.smallest;
 
-    final x = width / 2;
-    final y = height / 2;
+    final x = width! / 2;
+    final y = height! / 2;
     transformedPath = [
       ZMove.vector(ZVector.only(x: -x, y: -y)),
       ZLine.vector(ZVector.only(x: x, y: -y)),
@@ -82,7 +82,7 @@ class RenderZToBoxAdapter extends RenderZBox
 
     double sortValueTotal = 0;
     for (var i = 0; i < pointCount; i++) {
-      sortValueTotal += this.transformedPath[i].endRenderPoint.z;
+      sortValueTotal += this.transformedPath[i].endRenderPoint!.z!;
     }
     this.sortValue = sortValueTotal / pointCount;
   }
@@ -95,11 +95,11 @@ class RenderZToBoxAdapter extends RenderZBox
     Matrix4 matrix = Matrix4.translationValues(0, 0, 0);
     anchorParentData.transforms.forEach((transform) {
       final matrix4 = Matrix4.translationValues(
-          transform.translate.x, transform.translate.y, transform.translate.z);
+          transform.translate.x!, transform.translate.y!, transform.translate.z!);
 
-      matrix4.rotateX(transform.rotate.x);
-      matrix4.rotateY(-transform.rotate.y);
-      matrix4.rotateZ(transform.rotate.z);
+      matrix4.rotateX(transform.rotate.x!);
+      matrix4.rotateY(-transform.rotate.y!);
+      matrix4.rotateZ(transform.rotate.z!);
 
       matrix4.scale(transform.scale.x, transform.scale.y, transform.scale.z);
       matrix..multiply(matrix4);
@@ -127,7 +127,7 @@ class RenderZToBoxAdapter extends RenderZBox
       context.pushLayer(
         layer,
         (context, _) {
-          context.paintChild(child, offset - Offset(width / 2, height / 2));
+          context.paintChild(child!, offset - Offset(width! / 2, height! / 2));
         },
         Offset.zero,
         childPaintBounds: context.estimatedBounds,
