@@ -8,21 +8,21 @@ import '../path_command.dart';
 import '../renderer.dart';
 
 class RenderZShape extends RenderZBox {
-  Color _color;
+  Color? _color;
 
-  Color get color => _color;
+  Color? get color => _color;
 
-  set color(Color value) {
+  set color(Color? value) {
     if (_color == value) return;
     _color = value;
     markNeedsPaint();
   }
 
-  Color _backfaceColor;
+  Color? _backfaceColor;
 
-  Color get backfaceColor => _backfaceColor;
+  Color? get backfaceColor => _backfaceColor;
 
-  set backfaceColor(Color value) {
+  set backfaceColor(Color? value) {
     if (_backfaceColor == value) return;
     _backfaceColor = value;
     markNeedsPaint();
@@ -60,21 +60,21 @@ class RenderZShape extends RenderZBox {
     markNeedsLayout();
   }
 
-  List<ZPathCommand> _path;
+  List<ZPathCommand>? _path;
 
-  List<ZPathCommand> get path => _path;
+  List<ZPathCommand>? get path => _path;
 
-  set path(List<ZPathCommand> value) {
+  set path(List<ZPathCommand>? value) {
     if (_path == value) return;
     _path = value;
 
     markNeedsLayout();
   }
 
-  double _sortValue;
-  double get sortValue => _sortValue;
+  double? _sortValue;
+  double? get sortValue => _sortValue;
 
-  set sortValue(double value) {
+  set sortValue(double? value) {
     if (_sortValue == value) return;
     _sortValue = value;
   }
@@ -88,24 +88,24 @@ class RenderZShape extends RenderZBox {
     _visible = value;
   }
 
-  double _stroke;
+  double? _stroke;
 
-  double get stroke => _stroke;
+  double? get stroke => _stroke;
 
-  set stroke(double value) {
+  set stroke(double? value) {
     assert(value != null && value >= 0);
     if (_stroke == value) return;
     _stroke = value;
   }
 
   RenderZShape({
-    Color color,
-    Color backfaceColor,
+    Color? color,
+    Color? backfaceColor,
     ZVector front = const ZVector.only(z: 1),
     bool close = false,
     bool visible = true,
     bool fill = false,
-    double stroke = 1,
+    double? stroke = 1,
     List<ZPathCommand> path = const [],
   })  : assert(path != null),
         assert(front != null),
@@ -127,8 +127,8 @@ class RenderZShape extends RenderZBox {
   /// With this markNeedsPaint will only repaint this core object and not their ancestors
   bool get isRepaintBoundary => true;
 
-  ZVector _transformedFront;
-  ZVector normalVector;
+  late ZVector _transformedFront;
+  late ZVector normalVector;
   final Matrix4 matrix4 = Matrix4.identity();
 
   @override
@@ -152,7 +152,7 @@ class RenderZShape extends RenderZBox {
     normalVector = origin - _transformedFront;
     transformedPath = path;
     anchorParentData.transforms.reversed.forEach((matrix4) {
-      transformedPath = transformedPath
+      transformedPath = transformedPath!
           .map((e) =>
               e.transform(matrix4.translate, matrix4.rotate, matrix4.scale))
           .toList();
@@ -162,19 +162,19 @@ class RenderZShape extends RenderZBox {
     performSort();
   }
 
-  List<ZPathCommand> transformedPath = [];
+  List<ZPathCommand>? transformedPath = [];
 
   void performPathCommands() {
-    ZVector previousPoint = origin;
-    if (transformedPath.isEmpty) {
-      transformedPath.add(ZMove.vector(origin));
+    ZVector? previousPoint = origin;
+    if (transformedPath!.isEmpty) {
+      transformedPath!.add(ZMove.vector(origin));
     } else {
-      final first = transformedPath.first;
+      final first = transformedPath!.first;
       //Todo: Check this, I think not needed and can cause error
       if (!(first is ZMove)) {
-        transformedPath[0] = ZMove.vector(first.point());
+        transformedPath![0] = ZMove.vector(first.point()!);
       }
-      transformedPath.forEach((it) {
+      transformedPath!.forEach((it) {
         it.previous = previousPoint;
         previousPoint = it.endRenderPoint;
       });
@@ -183,10 +183,10 @@ class RenderZShape extends RenderZBox {
 
   @override
   void performSort() {
-    assert(transformedPath.isNotEmpty);
-    var pointCount = this.transformedPath.length;
-    var firstPoint = this.transformedPath[0].endRenderPoint;
-    var lastPoint = this.transformedPath[pointCount - 1].endRenderPoint;
+    assert(transformedPath!.isNotEmpty);
+    var pointCount = this.transformedPath!.length;
+    var firstPoint = this.transformedPath![0].endRenderPoint;
+    var lastPoint = this.transformedPath![pointCount - 1].endRenderPoint;
     // ignore the final point if self closing shape
     var isSelfClosing = pointCount > 2 && firstPoint == lastPoint;
     if (isSelfClosing) {
@@ -195,7 +195,7 @@ class RenderZShape extends RenderZBox {
 
     double sortValueTotal = 0;
     for (var i = 0; i < pointCount; i++) {
-      sortValueTotal += this.transformedPath[i].endRenderPoint.z;
+      sortValueTotal += this.transformedPath![i].endRenderPoint!.z!;
     }
     this.sortValue = sortValueTotal / pointCount;
   }
@@ -203,7 +203,7 @@ class RenderZShape extends RenderZBox {
   bool isFacingBack = false;
   bool showBackFace = true;
 
-  Color get renderColor {
+  Color? get renderColor {
     final isBackFaceColor = backfaceColor != null && isFacingBack;
     return isBackFaceColor ? backfaceColor : color;
   }
@@ -215,22 +215,22 @@ class RenderZShape extends RenderZBox {
 
     final renderer = ZRenderer(context.canvas);
     render(renderer);
-    final length = path.length;
+    final length = path!.length;
     if (length <= 1) {
       paintDot(renderer);
     } else {
-      isFacingBack = normalVector.z > 0;
+      isFacingBack = normalVector.z! > 0;
       if (!showBackFace && isFacingBack) {
         return super.paint(context, offset);
       }
 
-      var isTwoPoints = transformedPath.length == 2 && (path[1] is ZLine);
+      var isTwoPoints = transformedPath!.length == 2 && (path![1] is ZLine);
       var isClosed = !isTwoPoints && _close == true;
       final color = renderColor;
 
-      renderer.renderPath(transformedPath, isClosed: isClosed);
-      if (stroke != null && stroke > 0) renderer.stroke(color, stroke);
-      if (fill == true) renderer.fill(color);
+      renderer.renderPath(transformedPath!, isClosed: isClosed);
+      if (stroke != null && stroke! > 0) renderer.stroke(color!, stroke!);
+      if (fill == true) renderer.fill(color!);
     }
 
     //  context.canvas.restore();
@@ -241,12 +241,12 @@ class RenderZShape extends RenderZBox {
     if (stroke == 0.0) {
       return;
     }
-    final color = renderColor;
+    final color = renderColor!;
 
-    final point = transformedPath.first?.endRenderPoint ?? origin;
+    final point = transformedPath!.first.endRenderPoint ?? origin;
     renderer.begin();
-    final radius = stroke / 2;
-    renderer.circle(point.x, point.y, radius);
+    final radius = stroke! / 2;
+    renderer.circle(point.x!, point.y!, radius);
     renderer.closePath();
     renderer.fill(color);
   }
