@@ -4,15 +4,15 @@ import 'dart:math' as math;
 
 class ZCone extends ZCircle {
   final double diameter;
-  final double length;
+  final double? length;
 
   ZCone({
     this.length,
-    Key key,
-    @required this.diameter,
-    Color color,
+    Key? key,
+    required this.diameter,
+    Color? color,
     bool closed = false,
-    Color backfaceColor,
+    Color? backfaceColor,
     double stroke = 1,
     bool fill = true,
     ZVector front = const ZVector.only(z: 1),
@@ -31,7 +31,7 @@ class ZCone extends ZCircle {
   RenderZCone createRenderObject(BuildContext context) {
     return RenderZCone(
         color: color,
-        path: path,
+        path: path!,
         stroke: stroke,
         close: closed,
         fill: fill,
@@ -39,7 +39,7 @@ class ZCone extends ZCircle {
         backfaceColor: backfaceColor,
         front: front,
         diameter: diameter,
-        length: length);
+        length: length!);
   }
 
   @override
@@ -58,11 +58,11 @@ class ZCone extends ZCircle {
 }
 
 class RenderZCone extends RenderZShape {
-  double _length;
+  double? _length;
 
-  double get length => _length;
+  double? get length => _length;
 
-  set length(double value) {
+  set length(double? value) {
     assert(value != null && value >= 0);
     if (_length == value) return;
     _length = value;
@@ -81,10 +81,10 @@ class RenderZCone extends RenderZShape {
   }
 
   RenderZCone({
-    double length,
-    double diameter,
-    Color color,
-    Color backfaceColor,
+    required double length,
+    required double diameter,
+    Color? color,
+    Color? backfaceColor,
     ZVector front = const ZVector.only(z: 1),
     bool close = false,
     bool visible = true,
@@ -108,7 +108,7 @@ class RenderZCone extends RenderZShape {
   ZVector tangentA = ZVector.zero;
   ZVector tangentB = ZVector.zero;
 
-  ZVector apex;
+  ZVector? apex;
 
   @override
   void performLayout() {
@@ -117,7 +117,7 @@ class RenderZCone extends RenderZShape {
 
     apex = ZVector.only(z: length);
     anchorParentData.transforms.reversed.forEach((matrix4) {
-      apex = apex.transform(matrix4.translate, matrix4.rotate, matrix4.scale);
+      apex = apex!.transform(matrix4.translate, matrix4.rotate, matrix4.scale);
     });
     super.performLayout();
   }
@@ -140,7 +140,7 @@ class RenderZCone extends RenderZShape {
     if (!visible) {
       return;
     }
-    renderApex = apex - origin;
+    renderApex = apex! - origin;
     final scale = normalVector.magnitude();
     final apexDistance = renderApex.magnitude2d();
     final normalDistance = normalVector.magnitude2d();
@@ -154,7 +154,7 @@ class RenderZCone extends RenderZShape {
       return;
     }
 
-    final apexAngle = (math.atan2(normalVector.y, normalVector.x) + tau / 2);
+    final apexAngle = (math.atan2(normalVector.y!, normalVector.x!) + tau / 2);
     final projectLength = apexDistance / eccen;
     final projectAngle = math.acos(radius / projectLength);
 
@@ -168,12 +168,12 @@ class RenderZCone extends RenderZShape {
 
     final path = [
       ZMove.vector(tangentA),
-      ZLine.vector(apex),
+      ZLine.vector(apex!),
       ZLine.vector(tangentB),
     ];
 
     renderer.renderPath(path);
-    if (stroke > 0) renderer.stroke(color, stroke);
-    if (fill) renderer.fill(color);
+    if (stroke! > 0) renderer.stroke(color!, stroke!);
+    if (fill) renderer.fill(color!);
   }
 }
