@@ -12,11 +12,11 @@ class ZHemisphere extends StatelessWidget {
 
   final double stroke;
 
-  final Color color;
-  final bool visible;
+  final Color? color;
+  final bool? visible;
 
   // final ZVector front;
-  final Color backfaceColor;
+  final Color? backfaceColor;
 
   //var front = ZVector.only(z: 1);
   ZHemisphere({
@@ -50,16 +50,16 @@ class ZHemisphere extends StatelessWidget {
 }
 
 class _ZCylinderMiddle extends ZShape {
-  final double diameter;
+  final double? diameter;
 
   _ZCylinderMiddle(
-      {this.diameter, List<ZPathCommand> path, double stroke = 1, Color color})
+      {this.diameter, List<ZPathCommand>? path, double stroke = 1, Color? color})
       : super(path: [], stroke: stroke, color: color);
 
   @override
   _RenderZHemisphere createRenderObject(BuildContext context) {
     return _RenderZHemisphere(
-      path: path,
+      path: path!,
       stroke: stroke,
       diameter: diameter,
       color: color,
@@ -77,27 +77,27 @@ class _ZCylinderMiddle extends ZShape {
 }
 
 class _RenderZHemisphere extends RenderZShape {
-  double _diameter;
+  double? _diameter;
 
-  double get diameter => _diameter;
+  double? get diameter => _diameter;
 
-  set diameter(double value) {
+  set diameter(double? value) {
     if (_diameter == value) return;
     _diameter = value;
 
     markNeedsPaint();
   }
 
-  ZVector apex;
+  ZVector? apex;
   @override
   void performLayout() {
     final ZParentData anchorParentData = parentData as ZParentData;
     matrix4.setIdentity();
     // print('relayout ${anchorParentData.transforms.length}');
-    apex = ZVector.only(z: diameter / 2);
+    apex = ZVector.only(z: diameter! / 2);
     anchorParentData.transforms.reversed.forEach((matrix4) {
       //   print(matrix4);
-      apex = apex.transform(matrix4.translate, matrix4.rotate, matrix4.scale);
+      apex = apex!.transform(matrix4.translate, matrix4.rotate, matrix4.scale);
     });
     super.performLayout();
   }
@@ -109,16 +109,16 @@ class _RenderZHemisphere extends RenderZShape {
   }
 
   _RenderZHemisphere(
-      {List<ZPathCommand> path, double diameter, double stroke, Color color})
+      {required List<ZPathCommand> path, double? diameter, double? stroke, Color? color})
       : _diameter = diameter,
         super(path: path, stroke: stroke, color: color, fill: true);
 
   @override
   void render(ZRenderer renderer) {
-    final contourAngle = math.atan2(normalVector.y, normalVector.x);
-    final demoRadius = diameter / 2 * normalVector.magnitude();
-    final x = origin.x;
-    final y = origin.y;
+    final contourAngle = math.atan2(normalVector.y!, normalVector.x!);
+    final demoRadius = diameter! / 2 * normalVector.magnitude();
+    final x = origin.x!;
+    final y = origin.y!;
 
     final startAngle = contourAngle + tau / 4;
     final endAnchor = contourAngle - tau / 4;
@@ -127,7 +127,7 @@ class _RenderZHemisphere extends RenderZShape {
     renderer.move(origin);
     renderer.arc(x, y, demoRadius, startAngle, endAnchor);
     renderer.closePath();
-    if (stroke > 0) renderer.stroke(color, stroke);
-    if (fill) renderer.fill(color);
+    if (stroke! > 0) renderer.stroke(color!, stroke!);
+    if (fill) renderer.fill(color!);
   }
 }
