@@ -6,7 +6,7 @@ import 'package:flutter/physics.dart';
 import 'package:zflutter/zflutter.dart';
 
 class Dices extends StatefulWidget {
-  _PhysicsAnimation createState() => _DicesState();
+  _DicesState createState() => _DicesState();
 }
 
 class _DicesState extends State<Dices>
